# g2lserver/models/grammar_id.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from g2lserver.models.base_model_ import Model
from g2lserver import util


class GrammarID(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, id: str=None):  # noqa: E501
        """GrammarID - a model defined in Swagger

        :param id: The id of this GrammarID.  # noqa: E501
        :type id: str
        """
        self.swagger_types = {
            'id': str
        }

        self.attribute_map = {
            'id': 'id'
        }

        self._id = id

    @classmethod
    def from_dict(cls, dikt) -> 'GrammarID':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The GrammarID of this GrammarID.  # noqa: E501
        :rtype: GrammarID
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> str:
        """Gets the id of this GrammarID.


        :return: The id of this GrammarID.
        :rtype: str
        """
        return self._id

    @id.setter
    def id(self, id: str):
        """Sets the id of this GrammarID.


        :param id: The id of this GrammarID.
        :type id: str
        """
        if id is None:
            raise ValueError("Invalid value for `id`, must not be `None`")  # noqa: E501

        self._id = id