# g2lserver/models/pipelines.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from g2lserver.models.base_model_ import Model
from g2lserver import util


class Pipelines(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self):  # noqa: E501
        """Pipelines - a model defined in Swagger

        """
        self.swagger_types = {
        }

        self.attribute_map = {
        }

    @classmethod
    def from_dict(cls, dikt) -> 'Pipelines':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Pipelines of this Pipelines.  # noqa: E501
        :rtype: Pipelines
        """
        return util.deserialize_model(dikt, cls)